Decimal is not json-serializable in basket line attributes

+++ gap/apps/options/forms.py
@@ -32,8 +32,8 @@ class QuoteCalcForm(forms.Form):
 
 
 class QuoteCustomSizeForm(forms.Form):
-    width = forms.DecimalField()
-    height = forms.DecimalField()
+    width = forms.IntegerField()
+    height = forms.IntegerField()
 
 
 class ArtworkDeleteForm(forms.Form):
